pyproject.toml
[project]
name = "rag-pdf-extraction"
version = "0.1.0"
description = "RAG-enabled agentic solution for PDF information extraction"
authors = [
    {name = "User",email = "user@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "langchain (>=1.2.6,<2.0.0)",
    "langchain-anthropic (>=1.3.1,<2.0.0)",
    "langchain-openai (>=1.1.7,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pypdf (>=6.6.0,<7.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "chromadb (>=1.4.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sentence-transformers (>=5.2.0,<6.0.0)",
    "ollama (>=0.6.1,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
